Move core functionality to the new delta-core package.
This was done to make a PluginAPI possible which @thegail is working
on. I have started rewriting the UI code now that I understand SwiftUI
more. I have made rewrites of some of the managers to be singletons
that are much more smartly written than before. I still need to make
a proper startup sequence which I will likely use a
for this time for syncing.