Rewrite input system and create in game settings
The client now automatically captures the cursor without
requiring the user to click when joining a server. I also
added an in game menu which is reached by hitting escape
and contains settings that can be changed while playing
such as fov. This menu also contains a disconnect button.
I also created the start of a key mapping system but that
is yet to be integrated into the configuration system.