Fix 'Back to game' button in in-game menu
Clicking the button would not update the controls settings
of the input delegate but using the keyboard shortcut would
because there were two Back to game buttons, an invisible one
with the keyboard shortcut set and another visible one and I
didn't update the action of both to include updating the input
delegate. I fixed it by combining them back into just one.