Fix light level mutation for unlit sections in ChunkLighting (and hopefully issue #69)
The code made a false assumption that if the block light for a section was present,
the sky light should be too and vice versa. However, this meant that when setting
the light level of a block in a section with only one type of lighting present,
the type of lighting that was present also got overwritten when creating the new
empty lighting for the non-present type. This only happened when setting the light
level of the non-present light