Fix chunk lighting update infinite loop
This has evaded me so many times, and it was just some arguments
the wrong way around (arguments to clamp). This prevented
setBlockLightLevel and setSkyLightLevel from working correctly
because they just set the light level to 0 every time. This in
turn created an infinite loop in performLightLevelIncrease. I'm
glad that this is finally fixed