Make GUIList.meshes 4.5x faster
Achieved this by optimizing TextMeshBuilder.build through font improvements
and GUIElementMesh & GUIQuad through more optimized array creation (using
unsafe APIs). The font improvement was adding a cache for the descriptors
of ASCII characters to provide a fast path that can be taken much of the
time (in English at least).