Rework TexturePalette animation system to improve perf and implement frame interpolation
To render a texture palette you now need to create a MetalTexturePalette from
it which handles a time buffer, array texture, and texture states buffer. The
array texture is unchanging (unlike the previous implementation) and instead the
texture states buffer tells the shader which texture frames to use (and gets
updated each frame). The time buffer stores the current time (since the start of
rendering) which is used to implement smooth interpolation of textures which
explicitly specify that they should be interpolated. The difference is probably
most clear if you look at the glowing parts of the logs on a campfire before and
after the changes