Refactor app loading, warning/error modals, StorageManager, file system operations, troubleshooting, updater and progress tracking
Also refactored a bunch of errors to use a new RichError API which I'm
loving. Error types just hold the message, and additional context is
supplied using immutable instance methods such as `.with` and `becauseOf`.

I'm a lot happier with app loading and progress tracking now; the progress bar is more
accurate (takes into account skipped steps), loading is more concise and uses a more
declarative TaskProgress API, and loading is separated out from DeltaClientApp into its
own view (LoadAndThen). File system operations should now be done with FileSystem instead
of Swift's built-in FileManager (because it's a better API and we can control the errors
it emits to improve error messages). The updater was very weirdly architected so I've
rewritten that now, it ended up not having any internal state once I was done, so I
made it an enum with static methods. Troubleshooting works the same as before but the
code is just a bit more concise now.