Refactor ConfigManager into ManagedConfig and use an environment object instead of a singleton (mostly working)
Doesn't work for editing config in-game yet (because GameView is terribly coded and has stuff
in its init which should be in onAppear), but it's already way cleaner for the settings views
and stuff. It also fixes some cases where ConfigManager had to use fatal errors (due to needing
to create the global singleton no matter what). These changes also got rid of StorageManager