Fix animated textures with more animation frames than texture frames (i.e. lava which bounces instead of cycles)
The MetalTexturePalette implementation assumed that the number of animation frames
was equal to the number of texture frames, but that's violated by textures such as
lava. I refactored a bit so that MetalTexturePalette can easily know how many texture
frames a texture has without duplicating logic and making textures even harder to
refactor (I can see quite a bit of stuff to improve already.