Don't start rendering until the time has been received (avoids annoying flash)
Also took the opportunity to refactor events a bit (basically just removing
unnecessary public memberwise inits which were bloating things).

We used to just wait for terrain to finish downloading, but that often happens
before the time is received, so there was a brief period of time where the world
was rendered as night regardless of what the actual time was. It was pretty
noticeable.