Render a sky plane when in the overworld and use exponential fog underwater (replicating GL_FOG_EXP)
Took me ages to realise that OpenGL actually just has a built-in fog feature. I spent quite
a while searching around the decompiled vanilla sources and shaders to find the exponential
fog equation to no avail (before realising that it's actually just a documented built-in
GL feature).

The sky plane is rendered exactly how it is in Vanilla (or so I hope).